wallets/requirements.txt
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
mysqlclient==2.2.0
uvicorn[standard]==0.23.2
